Cleanup synthesizer .nn  and add short circuit imports for gans
