+++ sdk/opendp/smartnoise/synthesizers/pytorch/__init__.py
@@ -1,0 +1,4 @@
+from .pytorch_synthesizer import PytorchDPSynthesizer
+
+
+__all__ = ["PytorchDPSynthesizer"]
