+++ sdk/opendp/smartnoise/synthesizers/pytorch/nn/_discriminator.py
@@ -1,0 +1,19 @@
+import torch.nn as nn
+
+
+class Discriminator(nn.Module):
+    def __init__(self, input_dim, wasserstein=False):
+        super(Discriminator, self).__init__()
+        self.model = nn.Sequential(
+            nn.Linear(input_dim, 2 * input_dim // 3),
+            nn.LeakyReLU(0.2),
+            nn.Linear(2 * input_dim // 3, input_dim // 3),
+            nn.LeakyReLU(0.2),
+            nn.Linear(input_dim // 3, 1),
+        )
+
+        if not wasserstein:
+            self.model.add_module("activation", nn.Sigmoid())
+
+    def forward(self, x):
+        return self.model(x)
